refactor: make Pagination total a long

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/resource/AbstractResource.java
@@ -285,7 +285,7 @@ public abstract class AbstractResource {
         return new ArrayList<>(list).subList(startIndex, lastIndex);
     }
 
-    protected Links computePaginationLinks(int totalElements, PaginationParam paginationParam) {
+    protected Links computePaginationLinks(long totalElements, PaginationParam paginationParam) {
         return PaginationLinks.computePaginationLinks(
             uriInfo.getRequestUri(),
             uriInfo.getQueryParameters(),
