+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/resource/api/ApiMembersResource.java
@@ -70,7 +70,7 @@ public class ApiMembersResource extends AbstractResource {
 
         return new MembersResponse()
             .data(membersSubList)
-            .pagination(PaginationInfo.computePaginationInfo(members.size(), membersSubList.size(), paginationParam))
+            .pagination(PaginationInfo.computePaginationInfo((long) members.size(), membersSubList.size(), paginationParam))
             .links(computePaginationLinks(members.size(), paginationParam));
     }
